+++ src/event/event_action.rs
@@ -47,7 +47,7 @@ impl EventAction {
         status.tabs[1].refresh_if_needed()
     }
 
-    pub fn resize(status: &mut Status, width: usize, height: usize) -> Result<()> {
+    pub fn resize(status: &mut Status, width: u16, height: u16) -> Result<()> {
         status.resize(width, height)
     }
 
@@ -458,7 +458,7 @@ impl EventAction {
         status.set_height_for_edit_mode(status.index, Menu::Nothing)?;
         status.tabs[status.index].menu_mode = Menu::Nothing;
         let len = status.menu.len(Menu::Nothing);
-        let height = status.second_window_height()?;
+        let height = status.second_window_height()? as usize;
         status.menu.window = ContentWindow::new(len, height);
         status.tabs[status.index].menu_mode = Menu::InputSimple(InputSimple::Sort);
         status.set_focus_from_mode();
